--- a/bootstrap/cmd/bootstrap-pigeon/bootstrap_pigeon.go
+++ b/bootstrap/cmd/bootstrap-pigeon/bootstrap_pigeon.go
@@ -2473,9 +2473,6 @@ type rule struct {
 	name        string
 	displayName string
 	expr        any
-
-	leader        bool
-	leftRecursive bool
 }
 
 type choiceExpr struct {
@@ -2693,11 +2690,6 @@ type Stats struct {
 	ChoiceAltCnt map[string]map[string]int
 }
 
-type ruleWithExpsStack struct {
-	rule   *rule
-	estack []any
-}
-
 type parser struct {
 	filename string
 	pt       savepoint
@@ -2720,7 +2712,7 @@ type parser struct {
 	// variables stack, map of label to value
 	vstack []map[string]any
 	// rule stack, allows identification of the current rule in errors
-	rstack []ruleWithExpsStack
+	rstack []*rule
 
 	// parse fail
 	maxFailPos            position
@@ -2774,27 +2766,15 @@ func (p *parser) popV() {
 }
 
 func (p *parser) pushRule(rule *rule) {
-	p.rstack = append(p.rstack, ruleWithExpsStack{rule, []any{}})
+	p.rstack = append(p.rstack, rule)
 }
 
 func (p *parser) popRule() {
 	p.rstack = p.rstack[:len(p.rstack)-1]
 }
 
-func (p *parser) pushExpr(expr any) {
-	if len(p.rstack) == 0 {
-		return
-	}
-	p.rstack[len(p.rstack)-1].estack = append(
-		p.rstack[len(p.rstack)-1].estack, expr)
-}
-
-func (p *parser) popExpr() {
-	if len(p.rstack) == 0 {
-		return
-	}
-	p.rstack[len(p.rstack)-1].estack = p.rstack[len(p.rstack)-1].estack[:len(
-		p.rstack[len(p.rstack)-1].estack)-1]
+func (p *parser) getRule() *rule {
+	return p.rstack[len(p.rstack)-1]
 }
 
 // push a recovery expression with its labels to the recoveryStack
@@ -2864,7 +2844,7 @@ func (p *parser) addErrAt(err error, pos position, expected []string) {
 		if buf.Len() > 0 {
 			buf.WriteString(": ")
 		}
-		rule := p.rstack[len(p.rstack)-1].rule
+		rule := p.getRule()
 		if rule.displayName != "" {
 			buf.WriteString("rule " + rule.displayName)
 		} else {
@@ -3157,7 +3137,6 @@ func (p *parser) parseExpr(expr any) (any, bool) {
 		panic(errMaxExprCnt)
 	}
 
-	p.pushExpr(expr)
 	var val any
 	var ok bool
 	switch expr := expr.(type) {
@@ -3200,7 +3179,6 @@ func (p *parser) parseExpr(expr any) (any, bool) {
 	default:
 		panic(fmt.Sprintf("unknown expression type %T", expr))
 	}
-	p.popExpr()
 	return val, ok
 }
 
@@ -3344,7 +3322,7 @@ func (p *parser) parseCharClassMatcher(chr *charClassMatcher) (any, bool) {
 }
 
 func (p *parser) incChoiceAltCnt(ch *choiceExpr, altI int) {
-	choiceIdent := fmt.Sprintf("%s %d:%d", p.rstack[len(p.rstack)-1].rule.name, ch.pos.line, ch.pos.col)
+	choiceIdent := fmt.Sprintf("%s %d:%d", p.getRule().name, ch.pos.line, ch.pos.col)
 	m := p.ChoiceAltCnt[choiceIdent]
 	if m == nil {
 		m = make(map[string]int)