… builder/generated_static_code.go
… builder/left_recursion.go
… builder/left_recursion_test.go
… builder/scc.go
… builder/scc_test.go
… builder/static_code.go
… doc.go
… examples/calculator/calculator.go
… examples/indentation/indentation.go
… examples/json/json.go
… examples/json/optimized-grammar/json.go
… examples/json/optimized/json.go
… go.mod
… go.sum
… main.go
… pigeon.go
… targeted_test.go
… test/alternate_entrypoint/altentry.go
… test/andnot/andnot.go
… test/emptystate/emptystate.go
… test/errorpos/errorpos.go
… test/global_store/global_store.go
… test/goto/goto.go
… test/goto_state/goto_state.go
… test/issue_1/issue_1.go
… test/issue_18/issue_18.go
… test/issue_65/issue_65.go
… test/issue_65/optimized-grammar/issue_65.go
… test/issue_65/optimized/issue_65.go
… test/issue_70/issue_70.go