This PR is the first step towards the implementation of https://github.com/mna/pigeon/issues/120.
Implementation details are peeped here https://github.com/we-like-parsers/pegen_experiments/tree/master/pegen

Fixes: #120 
It also fixes https://github.com/mna/pigeon/issues/79