@sc07kvm 2023-06-25
`Can you please provide a general summary, how the left recursion handling is implemented.`

For the left recursion rule, we apply it in a loop, while memoizing the result of its last execution. For the first time, we consider that the rule was not executed successfully. We stop when we cannot get a longer result.

It is unlikely that I can describe it better than Guido van Rossum:
- https://medium.com/@gvanrossum_83706/left-recursive-peg-grammars-65dab3c580e1
- https://www.youtube.com/watch?v=QppWTvh7_sI&t=1659s

Priority support is implemented by saving rules and expressions in the stack. Therefore, at the stage of the new choice, we can check what choice we made last time.