@sc07kvm 2023-06-25
`Why do all the generated parsers change even if the flag support-left-recursion is not set to true? Can you please explain the changes.`

I tweaked the code a bit, now there are fewer changes. The remaining changes are needed to more conveniently integrate support for left recursion.

Major change:
```
func (p *parser) parseRule(rule *rule) (any, bool) {
	// debug code
	// memoize code
	// parse rule
	// debug code
	// memoize code
	return val, ok
}
```
->
```
func (p *parser) parseRuleWrap(rule *rule) (any, bool) {
	// debug code
	// choosing the right method
	if ... {
		val, ok := p.parseRule(rule)
	} else if ... {
		val, ok := p.parseRuleMemoize(rule)
	} else if ... {
		...
	}
	// debug code
	return val, ok
}

func (p *parser) parseRule(rule *rule) (any, bool) {
	// parse rule
}

func (p *parser) parseRuleMemoize(rule *rule) (any, bool) {
	// memoize code
	val, ok := p.parseRule(rule)
	// memoize code
	return val, ok
}
```