@sc07kvm 2023-06-26
> How does this change impact the performance of the generated parsers (optimized and not optimized). A good example for this is the json parser in examples/json.

Performance decreased by ~2% on **examples/json**

- Current master
https://github.com/mna/pigeon/tree/b0211796baa821ab4d73004fe458e5ae383b6a76
`go test -bench=. -benchmem ./examples/json`
```
goos: darwin
goarch: arm64
pkg: github.com/mna/pigeon/examples/json
BenchmarkPigeonJSONNoMemo-10                         157           7561956 ns/op         5003014 B/op     139782 allocs/op
BenchmarkPigeonJSONMemo-10                            49          24026827 ns/op        26645190 B/op     184036 allocs/op
BenchmarkPigeonJSONOptimized-10                      447           2694039 ns/op         2770617 B/op      69738 allocs/op
BenchmarkPigeonJSONOptimizedGrammar-10               164           7232733 ns/op         4998874 B/op     139647 allocs/op
BenchmarkStdlibJSON-10                              8139            138571 ns/op           74200 B/op       1010 allocs/op
PASS
ok      github.com/mna/pigeon/examples/json     7.893s
```

- this PR
https://github.com/sc07kvm/pigeon/tree/8a83e813e07d301e49df0453187acc51fa14ecad
`go test -bench=. -benchmem ./examples/json `
```
goos: darwin
goarch: arm64
pkg: github.com/mna/pigeon/examples/json
BenchmarkPigeonJSONNoMemo-10                         153           7679382 ns/op         5003028 B/op     139781 allocs/op
BenchmarkPigeonJSONMemo-10                            45          24330618 ns/op        26642812 B/op     184029 allocs/op
BenchmarkPigeonJSONOptimized-10                      439           2711443 ns/op         2770550 B/op      69737 allocs/op
BenchmarkPigeonJSONOptimizedGrammar-10               162           7397897 ns/op         4998891 B/op     139647 allocs/op
BenchmarkStdlibJSON-10                              8037            139898 ns/op           74205 B/op       1010 allocs/op
PASS
ok      github.com/mna/pigeon/examples/json     8.478s
```