@sc07kvm 2023-06-26
> Can you please check how the documentation would need to be changed (e.g. README.md, doc.go)

Added to doc.go, main.go