@breml 2023-06-26
> For the left recursion rule, we apply it in a loop, while memoizing the result of its last execution. For the first time, we consider that the rule was not executed successfully. We stop when we cannot get a longer result.

If I understand this correctly, left recursion support is only possible, if memoization is available. Of so, the flags `-optimize-parser` and `-support-left-recursion` are not compatible since with `-optimize-parser` the support for memoization is removed. Is this correct?