@sc07kvm 2023-06-28
> > For the left recursion rule, we apply it in a loop, while memoizing the result of its last execution. For the first time, we consider that the rule was not executed successfully. We stop when we cannot get a longer result.
> 
> If I understand this correctly, left recursion support is only possible, if memoization is available. Of so, the flags `-optimize-parser` and `-support-left-recursion` are not compatible since with `-optimize-parser` the support for memoization is removed. Is this correct?

Conceptually yes, memoization is needed.

But if the `-optimize-parser`, `-support-left-recursion` parameters are used and there is left recursion in the grammar, then the code responsible for memoization is added to the parser and memoization is applied only on one rule from each cycle in the grammar.

I added a test for this case:  **test/left_recursion/optimized**