@sc07kvm 2023-07-06
> An other angle, that is worrying me a little bit is, that at the moment we do not have grammar examples combining left recursion with some of the other features like throw-recover and state so it is currently unknown, if this implementation does behave as expected in these cases.

Added:
- ./test/left_recursion_state/ - modified ./test/left_recursion/ in the likeness of ./test/state
- ./test/left_recursion_labeled_failures/ - modified /test/labeled_failures/
- ./test/left_recursion_thrownrecover/ - modified ./test/thrownrecover/